Move to compiled babel
