+++ package.json
@@ -4,7 +4,9 @@
   "private": true,
   "main": "babel-node bin/server server",
   "scripts": {
-    "test": "test/runtests.sh"
+    "compile": "babel-compile -c tc-rules/babel src:lib test:.test",
+    "test":    "mocha .test/*_test.js",
+    "pretest": "npm run compile"
   },
   "repository": {
     "type": "git",
@@ -23,7 +25,16 @@
     "taskcluster-base": "^0.12.1",
     "taskcluster-client": "^0.22.6",
     "taskcluster-lib-loader": "^0.1.1",
-    "taskcluster-lib-testing": "^0.11.1"
+    "taskcluster-lib-testing": "^0.11.1",
+    "babel-runtime":        "^6.6.1",
+    "babel-compile":        "^1.0.2",
+    "tc-rules":             "^5.0.0",
+    "babel-preset-es2015":  "^6.6.0",
+    "babel-plugin-syntax-async-functions":          "^6.5.0",
+    "babel-plugin-transform-async-to-generator":    "^6.5.0",
+    "babel-plugin-transform-runtime":               "^6.6.0",
+    "babel-plugin-transform-strict-mode":           "^6.5.2",
+    "babel-plugin-syntax-trailing-function-commas": "^6.5.0"
   },
   "devDependencies": {
     "mocha": "2.1.0"
